requirements.txt
# Django CMS Requirements
Django>=5.0,<6.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.0
django-ckeditor-5>=0.2.10
pyotp>=2.9.0
qrcode>=7.4
Pillow>=10.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
reportlab>=4.0.0
pypdf>=4.0.0
pdfplumber>=0.10.0
django-allauth>=0.60.0
